pyproject.toml
[project]
name = "django-rest-api-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "rodrigo.resque",email = "rodrigo.resque@tjpa.jus.br"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pygments (>=2.19.2,<3.0.0)"
]

[tool.poetry]
packages = [{include = "django_rest_api_tutorial", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
